Renames TypeEnv to Scope
Renames Scope to Strategy

Initializes TypeEnv to represent both locals and globals
