+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/typer/Strategy.kt
@@ -1,0 +1,14 @@
+package org.partiql.planner.internal.typer
+
+/**
+ * Variable resolution strategies — https://partiql.org/assets/PartiQL-Specification.pdf#page=35
+ *
+ * | Value      | Strategy              | Scoping Rules |
+ * |------------+-----------------------+---------------|
+ * | LOCAL      | local-first lookup    | Rules 1, 2    |
+ * | GLOBAL     | global-first lookup   | Rule 3        |
+ */
+internal enum class Strategy {
+    LOCAL,
+    GLOBAL,
+}
